chore(www): move skeleton cards to marketing and tidy homepage
Relocate skeleton-cards into the marketing module, remove stale EXPLORATION.md, and simplify homepage/cards-grid structure ahead of broader cleanup.
